notebooks/bert_topic/top2vec_requirements.txt
absl-py==1.4.0
alabaster==0.7.13
alembic==1.10.3
anyascii==0.3.2
anyio==3.6.2
astatine==0.3.2
astor==0.8.1
astpretty==3.0.0
astroid==2.15.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autoflake==1.7.8
Babel==2.12.1
bandit==1.7.5
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmaes==0.9.1
cognitive-complexity==1.3.0
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
contractions==0.1.73
coverage==7.2.3
cycler==0.11.0
Cython @ file:///D:/bld/cython_1680712482544/work
darglint==1.8.1
distlib==0.3.6
dlint==0.14.0
doc8==1.1.1
docformatter==1.6.0
docutils==0.19
domdf-python-tools==3.6.0
eradicate==2.2.0
exceptiongroup==1.1.1
fastapi==0.95.0
filelock==3.11.0
flake8==5.0.4
flake8-2020==1.7.0
flake8-aaa==0.14.0
flake8-annotations==3.0.0
flake8-annotations-complexity==0.0.7
flake8-annotations-coverage==0.0.6
flake8-bandit==4.1.1
flake8-black==0.3.6
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.6.0
flake8-bugbear==23.3.12
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-cognitive-complexity==0.1.0
flake8-comments==0.1.2
flake8-comprehensions==3.11.1
flake8-debugger==4.1.2
flake8-django==1.1.5
flake8-docstrings==1.7.0
flake8-encodings==0.5.0.post1
flake8-eradicate==1.4.0
flake8-executable==2.1.3
flake8-expression-complexity==0.0.11
flake8-fastapi==0.7.0
flake8-fixme==1.1.1
flake8-functions==0.0.7
flake8-functions-names==0.4.0
flake8-future-annotations==0.0.5
flake8-helper==0.2.1
flake8-isort==6.0.0
flake8-literal==1.3.0
flake8-logging-format==0.9.0
flake8-markdown==0.4.0
flake8-mutable==1.2.0
flake8-no-pep420==2.3.0
flake8-noqa==1.3.1
flake8-pie==0.16.0
flake8-plugin-utils==1.3.2
flake8-pyi==22.11.0
flake8-pylint==0.2.0
flake8-pytest-style==1.7.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-secure-coding-standard==1.4.0
flake8-string-format==0.3.0
flake8-tidy-imports==4.8.0
flake8-typing-imports==1.14.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.4
flake8-variables-names==0.0.5
flake8-warnings==0.4.0
flake8_simplify==0.20.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
h5py==3.8.0
hdbscan @ file:///D:/bld/hdbscan_1675624117633/work
hypothesis==6.70.2
hypothesmith==0.1.9
idna==3.4
imagesize==1.4.1
importlib-metadata==6.3.0
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
jupyterlab-flake8==0.7.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lark-parser==0.12.0
lazy-object-proxy==1.9.0
libclang==16.0.0
libcst==0.4.9
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
mr-proper==0.0.7
mypy-extensions==1.0.0
natsort==8.3.1
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
optuna==3.1.1
packaging==23.0
pandas==2.0.0
pandas-vet==0.2.3
pathspec==0.10.3
pbr==5.11.1
pep8-naming==0.13.3
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
protobuf==3.19.6
pyahocorasick==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybetter==0.4.1
pycln==2.1.3
pycodestyle==2.9.1
pydantic==1.10.7
pydocstyle==6.3.0
pyemojify==0.2.0
pyflakes==2.5.0
Pygments==2.15.0
pylint==2.17.2
pynndescent==0.5.8
pyparsing==3.0.9
pyproject_api==1.5.1
pytest==7.2.2
pytest-cov==4.0.0
pytest-sugar==0.9.6
python-dateutil==2.8.2
python-dev-tools==2023.3.24
pytz==2023.3
pyupgrade==3.3.1
PyYAML==6.0
regex==2023.3.23
removestar==1.3.1
requests==2.28.2
requests-oauthlib==1.3.1
restructuredtext-lint==1.4.0
rich==13.3.3
rsa==4.9
scikit-learn @ file:///D:/bld/scikit-learn_1679676068685/work
scipy==1.9.3
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.9
ssort==0.11.6
starlette==0.26.1
stdlib-list==0.8.0
stevedore==5.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.10.0
termcolor==2.2.0
textsearch==0.0.24
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tokenize-rt==5.0.0
tomli==2.0.1
tomlkit==0.11.7
top2vec==1.0.29
tox==4.4.11
tox-travis==0.12
tqdm==4.65.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
umap-learn==0.5.3
untokenize==0.1.1
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.2.3
wincertstore==0.2
wordcloud==1.8.2.2
wrapt==1.15.0
zipp==3.15.0
